parasite: fix size of cmd
cmd is declarated in parasite-head-x86-64.S
__export_parasite_cmd:
        .long 0

Its size is 4 bytes, so it should be int instead of long.

This patch fixes a bug in parasite execute,
where a part of args are overwritten by cmd.

memcpy(ctl->addr_cmd, &cmd, sizeof(cmd));

Reported-by: Stanislav Kinsbursky <skinsbursky@parallels.com>
Acked-by: Cyrill Gorcunov <gorcunov@openvz.org>

+++ parasite-syscall.c
@@ -221,7 +221,7 @@ err:
 	return ret;
 }
 
-static int parasite_execute_by_pid(unsigned long cmd, struct parasite_ctl *ctl,
+static int parasite_execute_by_pid(unsigned int cmd, struct parasite_ctl *ctl,
 			    pid_t pid,
 			    void *args, int args_size)
 {
@@ -265,7 +265,7 @@ static int parasite_execute_by_pid(unsigned long cmd, struct parasite_ctl *ctl,
 	return ret;
 }
 
-static int parasite_execute(unsigned long cmd, struct parasite_ctl *ctl,
+static int parasite_execute(unsigned int cmd, struct parasite_ctl *ctl,
 			    void *args, int args_size)
 {
 	return parasite_execute_by_pid(cmd, ctl, ctl->pid, args, args_size);
